support Elm 0.19
