+++ .gitignore
@@ -1,4 +1,5 @@
 .vscode
 node_modules
 elm-stuff
-elm.js
+src/elm.js
+src/old.js
