+++ bin/elm-outdated
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require("../index.js");
+require("../src/index.js");
