+++ elm.json
@@ -1,0 +1,23 @@
+{
+    "type": "application",
+    "source-directories": [
+        "./src"
+    ],
+    "elm-version": "0.19.0",
+    "dependencies": {
+        "direct": {
+            "elm/core": "1.0.0",
+            "elm/json": "1.0.0",
+            "elm/project-metadata-utils": "1.0.0",
+            "elm/random": "1.0.0"
+        },
+        "indirect": {
+            "elm/parser": "1.0.0",
+            "elm/time": "1.0.0"
+        }
+    },
+    "test-dependencies": {
+        "direct": {},
+        "indirect": {}
+    }
+}
